Release candidate 1

+++ app/src/main/java/com/bartoszdrozd/mediapp/healthforms/dtos/HeartFormDTO.kt
@@ -1,12 +1,12 @@
 package com.bartoszdrozd.mediapp.healthforms.dtos
 
 data class HeartFormDTO(
-    var restingBloodPressure: Int? = -1,
-    var serumCholesterol: Int? = -1,
-    var fastingBloodSugar: Int? = -1,
+    var restingBloodPressure: Float? = -1f,
+    var serumCholesterol: Float? = -1f,
+    var fastingBloodSugar: Float? = -1f,
     var restingECG: Int = -1,
-    var maxHR: Int? = -1,
-    var stDepression: Int? = -1,
+    var maxHR: Float? = -1f,
+    var stDepression: Float? = -1f,
     var chestPainType: Int = 0,
     var exerciseInducedAngina: Int = 0,
     var peakSTSegment: Int = 0,
